test(Anchor): migrate unit tests from Test.Ursa to headless project
